fix(thermostat-led-display): proper events handling
There were a couple of bugs like data on display being updated late (not instantly) and the relay to be ON when the schedule is over. This was due to some small bugs in the vents handling.
I worked on a doc to map all events and I will commit it in one of the next commits together with a refactoring of events names.